5/31

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: nsakanou <nsakanou@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/17 11:44:30 by nsakanou          #+#    #+#             */
-/*   Updated: 2023/05/26 09:47:06 by nsakanou         ###   ########.fr       */
+/*   Updated: 2023/05/31 17:06:32 by nsakanou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
